Don't bother defining Symbol#to_proc
Turns out there are some diffrences in how ruby treats method(*args) and proc {|*args|} (and proc {|*args|} is valid in ruby 1.8.6, which for some reason I didn't think it was). But now I can't be bothered to support it, most people are running a ruby with native support anyway.